revert changes from prev design iteration
