+++ catalog/app/utils/BucketConfig.ts
@@ -54,9 +54,3 @@ export function useIsInStack() {
     [bucketConfigs],
   )
 }
-
-export function useBucketConfig(bucket: string) {
-  // FIXME: load only bucket
-  const bucketConfigs = useBucketConfigs()
-  return bucketConfigs.find(({ name }) => name === bucket)
-}
